fixed the spacing for the nodeMCU pins
moved the LEDs to the bottom of the display

Changed the current setting resistor for both the LEDs and the buzzer to 100 ohms.

remapped the rfid card reader to use pins that can also be used as hardware spi pins.

Put both the buzzer and the LEDs on pwm compatible pins

replaced the 6pin jst connector with an 8pin connector.

added the other hardware spi pins to the jst connector.